fixed couple of test and reverted ownableupgradeabl
