+++ packages/contracts/contracts/CollSurplusPool.sol
@@ -2,7 +2,7 @@
 
 pragma solidity 0.8.11;
 
-import "@openzeppelin/contracts-upgradeable/access/OwnableUpgradeable.sol";
+// import "@openzeppelin/contracts-upgradeable/access/OwnableUpgradeable.sol";
 
 import "./Interfaces/ICollSurplusPool.sol";
 import "./Dependencies/SafeMath.sol";
@@ -11,7 +11,7 @@ import "./Dependencies/CheckContract.sol";
 import "./Dependencies/console.sol";
 
 
-contract CollSurplusPool is OwnableUpgradeable, CheckContract, ICollSurplusPool {
+contract CollSurplusPool is Ownable, CheckContract, ICollSurplusPool {
     using SafeMath for uint256;
 
     bool public isInitialized;
@@ -44,15 +44,15 @@ contract CollSurplusPool is OwnableUpgradeable, CheckContract, ICollSurplusPool 
     )
         external
         override
-        initializer
+        onlyOwner
     {
-		require(!isInitialized, "Already initialized");
+		// require(!isInitialized, "Already initialized");
 		checkContract(_borrowerOperationsAddress);
 		checkContract(_troveManagerAddress);
 		checkContract(_activePoolAddress);
-		isInitialized = true;
+		// isInitialized = true;
 
-		__Ownable_init();
+		// __Ownable_init();
 
         borrowerOperationsAddress = _borrowerOperationsAddress;
         troveManagerAddress = _troveManagerAddress;
@@ -62,7 +62,7 @@ contract CollSurplusPool is OwnableUpgradeable, CheckContract, ICollSurplusPool 
         emit TroveManagerAddressChanged(_troveManagerAddress);
         emit ActivePoolAddressChanged(_activePoolAddress);
 
-        renounceOwnership();
+        _renounceOwnership();
     }
 
     /* Returns the ETH state variable at ActivePool address.
