+++ packages/contracts/contracts/Dependencies/Ownable.sol
@@ -14,53 +14,53 @@ pragma solidity 0.8.11;
  * `onlyOwner`, which can be applied to your functions to restrict their use to
  * the owner.
  */
-// contract Ownable {
-//     address private _owner;
+contract Ownable {
+    address private _owner;
 
-//     event OwnershipTransferred(address indexed previousOwner, address indexed newOwner);
+    event OwnershipTransferred(address indexed previousOwner, address indexed newOwner);
 
-//     /**
-//      * @dev Initializes the contract setting the deployer as the initial owner.
-//      */
-//     constructor () {
-//         _owner = msg.sender;
-//         emit OwnershipTransferred(address(0), msg.sender);
-//     }
+    /**
+     * @dev Initializes the contract setting the deployer as the initial owner.
+     */
+    constructor () {
+        _owner = msg.sender;
+        emit OwnershipTransferred(address(0), msg.sender);
+    }
 
-//     /**
-//      * @dev Returns the address of the current owner.
-//      */
-//     function owner() public view returns (address) {
-//         return _owner;
-//     }
+    /**
+     * @dev Returns the address of the current owner.
+     */
+    function owner() public view returns (address) {
+        return _owner;
+    }
 
-//     /**
-//      * @dev Throws if called by any account other than the owner.
-//      */
-//     modifier onlyOwner() {
-//         require(isOwner(), "Ownable: caller is not the owner");
-//         _;
-//     }
+    /**
+     * @dev Throws if called by any account other than the owner.
+     */
+    modifier onlyOwner() {
+        require(isOwner(), "Ownable: caller is not the owner");
+        _;
+    }
 
-//     /**
-//      * @dev Returns true if the caller is the current owner.
-//      */
-//     function isOwner() public view returns (bool) {
-//         return msg.sender == _owner;
-//     }
+    /**
+     * @dev Returns true if the caller is the current owner.
+     */
+    function isOwner() public view returns (bool) {
+        return msg.sender == _owner;
+    }
 
-//     /**
-//      * @dev Leaves the contract without owner. It will not be possible to call
-//      * `onlyOwner` functions anymore.
-//      *
-//      * NOTE: Renouncing ownership will leave the contract without an owner,
-//      * thereby removing any functionality that is only available to the owner.
-//      *
-//      * NOTE: This function is not safe, as it doesn’t check owner is calling it.
-//      * Make sure you check it before calling it.
-//      */
-//     function _renounceOwnership() internal {
-//         emit OwnershipTransferred(_owner, address(0));
-//         _owner = address(0);
-//     }
-// }
+    /**
+     * @dev Leaves the contract without owner. It will not be possible to call
+     * `onlyOwner` functions anymore.
+     *
+     * NOTE: Renouncing ownership will leave the contract without an owner,
+     * thereby removing any functionality that is only available to the owner.
+     *
+     * NOTE: This function is not safe, as it doesn’t check owner is calling it.
+     * Make sure you check it before calling it.
+     */
+    function _renounceOwnership() internal {
+        emit OwnershipTransferred(_owner, address(0));
+        _owner = address(0);
+    }
+}
