+++ packages/contracts/contracts/Interfaces/ILiquityBase.sol
@@ -3,14 +3,14 @@
 pragma solidity 0.8.11;
 
 import "./IPriceFeed.sol";
-import "./IKumoParameters.sol";
+// import "./IKumoParameters.sol";
 
 
 interface ILiquityBase {
     
     event VaultParametersBaseChanged(address indexed newAddress);
 
-	function kumoParams() external view returns (IKumoParameters);
+	// function kumoParams() external view returns (IKumoParameters);
 
     function priceFeed() external view returns (IPriceFeed);
 }
