+++ packages/contracts/test copy/ConnectContractsTest.js
